+++ PKVault.Backend/storage/data-action/DeletePkmVariantAction.cs
@@ -14,6 +14,13 @@ public class DeletePkmVariantAction(
         async Task<DataActionPayload> act(string pkmVariantId)
         {
             var pkmVariant = await pkmVariantLoader.GetEntity(pkmVariantId);
+            var dto = await pkmVariantLoader.CreateDTO(pkmVariant);
+
+            if (!dto.CanDelete)
+            {
+                throw new ArgumentException($"PkmVariant cannot be released: {pkmVariant.Id}");
+            }
+
             var pkm = await pkmVariantLoader.GetPKM(pkmVariant);
 
             await pkmVariantLoader.DeleteEntity(pkmVariant);
@@ -31,7 +38,7 @@ public class DeletePkmVariantAction(
 
             return new(
                 type: DataActionType.DELETE_PKM_VERSION,
-                parameters: [pkm.Nickname, pkmVariant.Generation]
+                parameters: [pkm.Nickname, pkmVariant.Context]
             );
         }
